backend/pyproject.toml
[project]
name = "nft-generative-art-backend"
version = "0.1.0"
description = "Backend service for NFT Generative Art with Provable Randomness"
requires-python = ">=3.10"
dependencies = [
    "web3>=6.0.0",
    "diffusers>=0.27.0",
    "torch>=2.0.0",
    "transformers>=4.30.0",
    "accelerate>=0.20.0",
    "safetensors>=0.4.0",
    "Pillow>=10.0.0",
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"
